Refactor PurchasesOrchestratorTests
