+++ Sources/Misc/StoreKitVersion.swift
@@ -33,7 +33,7 @@ public enum StoreKitVersion: Int {
 public extension StoreKitVersion {
 
     /// Let RevenueCat use the most appropiate version of StoreKit
-    static let `default` = Self.storeKit1
+    static let `default` = Self.storeKit2
 
 }
