+++ Tests/StoreKitUnitTests/PurchasesOrchestratorSK1Tests.swift
@@ -18,11 +18,13 @@ import StoreKit
 import XCTest
 
 @available(iOS 14.0, tvOS 14.0, macOS 11.0, watchOS 7.0, *)
-class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
+class PurchasesOrchestratorSK1Tests: BasePurchasesOrchestratorTests, PurchasesOrchestratorTests {
+
+    override class var storeKitVersion: StoreKitVersion { .storeKit1 }
 
     // MARK: - StoreFront Changes
 
-    func testClearCachedProductsAndOfferingsAfterStorefrontChangesWithSK1() async throws {
+    func testClearCachedProductsAndOfferingsAfterStorefrontChanges() async throws {
         self.orchestrator.storeKit1WrapperDidChangeStorefront(storeKit1Wrapper)
 
         expect(self.mockOfferingsManager.invokedInvalidateAndReFetchCachedOfferingsIfAppropiateCount) == 1
@@ -31,8 +33,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
 
     // MARK: - Purchasing
 
-    func testPurchaseSK1PackageSendsReceiptToBackendIfSuccessful() async throws {
-        self.customerInfoManager.stubbedCachedCustomerInfoResult = mockCustomerInfo
+    func testPurchasePostsReceipt() async throws {
         self.backend.stubbedPostReceiptResult = .success(mockCustomerInfo)
 
         let product = try await self.fetchSk1Product()
@@ -57,32 +58,147 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         expect(self.receiptFetcher.receiptDataReceivedRefreshPolicy) == .always
 
         expect(self.backend.invokedPostReceiptDataCount) == 1
+        expect(self.backend.invokedPostReceiptData).to(beTrue())
+        expect(self.backend.invokedPostReceiptDataParameters?.data).toNot(beNil())
         expect(self.backend.invokedPostReceiptDataParameters?.productData).toNot(beNil())
         expect(self.backend.invokedPostReceiptDataParameters?.transactionData.presentedOfferingID) == "offering"
         expect(self.backend.invokedPostReceiptDataParameters?.transactionData.source.initiationSource) == .purchase
     }
 
-    func testSK1PurchaseDoesNotAlwaysRefreshReceiptInProduction() async throws {
-        self.customerInfoManager.stubbedCachedCustomerInfoResult = self.mockCustomerInfo
+    func testPurchaseReturnsCorrectValues() async throws {
+        backend.stubbedPostReceiptResult = .success(mockCustomerInfo)
+
+        let product = try await self.fetchSk1Product()
+        let payment = storeKit1Wrapper.payment(with: product)
+
+        let (transaction, customerInfo, error, userCancelled) = await withCheckedContinuation { continuation in
+            orchestrator.purchase(sk1Product: product,
+                                  payment: payment,
+                                  package: nil,
+                                  wrapper: self.storeKit1Wrapper) { transaction, customerInfo, error, userCancelled in
+                continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
+            }
+        }
+
+        expect(transaction?.sk1Transaction?.productIdentifier) == product.productIdentifier
+        expect(userCancelled) == false
+        expect(error).to(beNil())
+
+        let expectedCustomerInfo: CustomerInfo = .emptyInfo
+        expect(customerInfo) == expectedCustomerInfo
+    }
+
+    func testPurchaseDoesNotPostReceiptIfPurchaseFailed() async throws {
+        storeKit1Wrapper.mockAddPaymentTransactionState = .failed
+        storeKit1Wrapper.mockTransactionError = NSError(domain: SKErrorDomain,
+                                                        code: SKError.Code.paymentCancelled.rawValue)
+
+        let product = try await self.fetchSk1Product()
+        let offer = PromotionalOffer.SignedData(identifier: "",
+                                                keyIdentifier: "",
+                                                nonce: UUID(),
+                                                signature: "",
+                                                timestamp: 0)
+
+        let (transaction, customerInfo, error, userCancelled) = await withCheckedContinuation { continuation in
+            orchestrator.purchase(sk1Product: product,
+                                  promotionalOffer: offer,
+                                  package: nil,
+                                  wrapper: self.storeKit1Wrapper) { transaction, customerInfo, error, userCancelled in
+                continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
+            }
+        }
+        expect(self.backend.invokedPostReceiptData) == false
+        expect(transaction?.sk1Transaction?.transactionState) == .failed
+        expect(customerInfo).to(beNil())
+        expect(error).to(matchError(ErrorCode.purchaseCancelledError))
+        expect(userCancelled) == true
+    }
+
+    func testPurchaseWithPromotionalOfferPostsReceiptIfSuccessful() async throws {
+        backend.stubbedPostReceiptResult = .success(mockCustomerInfo)
+
+        let product = try await fetchSk1Product()
+        let storeProduct = StoreProduct(sk1Product: product)
+
+        let offer = PromotionalOffer.SignedData(identifier: "",
+                                                keyIdentifier: "",
+                                                nonce: UUID(),
+                                                signature: "",
+                                                timestamp: 0)
+
+        _ = await withCheckedContinuation { continuation in
+            orchestrator.purchase(sk1Product: product,
+                                  promotionalOffer: offer,
+                                  package: nil,
+                                  wrapper: self.storeKit1Wrapper) { transaction, customerInfo, error, userCancelled in
+                continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
+            }
+        }
+
+        expect(self.backend.invokedPostReceiptDataCount) == 1
+        expect(self.backend.invokedPostReceiptDataParameters?.productData).toNot(beNil())
+    }
+
+    func testPurchaseWithInvalidPromotionalOfferSignatureFails() async throws {
+        storeKit1Wrapper.mockAddPaymentTransactionState = .failed
+        storeKit1Wrapper.mockTransactionError = NSError(domain: SKErrorDomain,
+                                                        code: SKError.Code.invalidSignature.rawValue)
+        let product = try await self.fetchSk1Product()
+        let offer = PromotionalOffer.SignedData(identifier: "",
+                                                keyIdentifier: "",
+                                                nonce: UUID(),
+                                                signature: "",
+                                                timestamp: 0)
+
+        let (transaction, customerInfo, error, userCancelled) = await withCheckedContinuation { continuation in
+            orchestrator.purchase(sk1Product: product,
+                                  promotionalOffer: offer,
+                                  package: nil,
+                                  wrapper: self.storeKit1Wrapper) { transaction, customerInfo, error, userCancelled in
+                continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
+            }
+        }
+        expect(transaction).toNot(beNil())
+        expect(customerInfo).to(beNil())
+        expect(error).to(matchError(ErrorCode.invalidPromotionalOfferError))
+        expect(userCancelled) == false
+        expect(self.backend.invokedPostReceiptData) == false
+    }
+
+    func testPurchaseCancelled() async throws {
+        storeKit1Wrapper.mockAddPaymentTransactionState = .failed
+        storeKit1Wrapper.mockTransactionError = NSError(domain: SKErrorDomain,
+                                                        code: SKError.Code.paymentCancelled.rawValue)
+        let product = try await self.fetchSk1Product()
+        let (transaction, customerInfo, error, userCancelled) = await withCheckedContinuation { continuation in
+            orchestrator.purchase(product: StoreProduct(sk1Product: product),
+                                  package: nil) { transaction, customerInfo, error, userCancelled in
+                continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
+            }
+        }
+        expect(transaction).toNot(beNil())
+        expect(customerInfo).to(beNil())
+        expect(error).to(matchError(ErrorCode.purchaseCancelledError))
+        expect(userCancelled) == true
+        expect(self.backend.invokedPostReceiptData) == false
+    }
+
+    // MARK: - Purchasing, StoreKit 1 only
+
+    func testPurchaseSK1DoesNotAlwaysRefreshReceiptInProduction() async throws {
         self.backend.stubbedPostReceiptResult = .success(self.mockCustomerInfo)
 
         self.systemInfo.stubbedIsSandbox = false
 
         let product = try await self.fetchSk1Product()
-        let storeProduct = try await self.fetchSk1StoreProduct()
-
-        let package = Package(identifier: "package",
-                              packageType: .monthly,
-                              storeProduct: .from(product: storeProduct),
-                              offeringIdentifier: "offering")
-
         let payment = self.storeKit1Wrapper.payment(with: product)
 
         _ = await withCheckedContinuation { continuation in
             self.orchestrator.purchase(
                 sk1Product: product,
                 payment: payment,
-                package: package,
+                package: nil,
                 wrapper: self.storeKit1Wrapper
             ) { transaction, customerInfo, error, userCancelled in
                 continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
@@ -93,7 +209,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         expect(self.receiptFetcher.receiptDataReceivedRefreshPolicy) == .onlyIfEmpty
     }
 
-    func testPurchaseSK1PackageRetriesReceiptFetchIfEnabled() async throws {
+    func testPurchaseSK1RetriesReceiptFetchIfEnabled() async throws {
         self.systemInfo = .init(
             platformInfo: nil,
             finishTransactions: false,
@@ -103,24 +219,16 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         )
         self.setUpStoreKit1Wrapper()
         self.setUpOrchestrator()
-        self.customerInfoManager.stubbedCachedCustomerInfoResult = self.mockCustomerInfo
         self.backend.stubbedPostReceiptResult = .success(self.mockCustomerInfo)
 
         let product = try await self.fetchSk1Product()
-        let storeProduct = StoreProduct(sk1Product: product)
-
-        let package = Package(identifier: "package",
-                              packageType: .monthly,
-                              storeProduct: storeProduct,
-                              offeringIdentifier: "offering")
-
         let payment = self.storeKit1Wrapper.payment(with: product)
 
         _ = await withCheckedContinuation { continuation in
             self.orchestrator.purchase(
                 sk1Product: product,
                 payment: payment,
-                package: package,
+                package: nil,
                 wrapper: self.storeKit1Wrapper
             ) { transaction, customerInfo, error, userCancelled in
                 continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
@@ -129,7 +237,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
 
         expect(self.receiptFetcher.receiptDataCalled) == true
         expect(self.receiptFetcher.receiptDataReceivedRefreshPolicy) == .retryUntilProductIsFound(
-            productIdentifier: storeProduct.productIdentifier,
+            productIdentifier: product.productIdentifier,
             maximumRetries: TransactionPoster.receiptRetryCount,
             sleepDuration: TransactionPoster.receiptRetrySleepDuration
         )
@@ -138,17 +246,10 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         expect(self.backend.invokedPostReceiptDataParameters?.productData).toNot(beNil())
     }
 
-    func testPurchaseSK1PackageWithNoProductIdentifierDoesNotPostReceipt() async throws {
-        self.customerInfoManager.stubbedCachedCustomerInfoResult = self.mockCustomerInfo
+    func testPurchaseSK1WithNoProductIdentifierDoesNotPostReceipt() async throws {
         self.backend.stubbedPostReceiptResult = .success(self.mockCustomerInfo)
 
         let product = try await self.fetchSk1Product()
-        let storeProduct = StoreProduct(sk1Product: product)
-        let package = Package(identifier: "package",
-                              packageType: .monthly,
-                              storeProduct: storeProduct,
-                              offeringIdentifier: "offering")
-
         let payment = self.storeKit1Wrapper.payment(with: product)
         payment.productIdentifier = ""
 
@@ -157,7 +258,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
             self.orchestrator.purchase(
                 sk1Product: product,
                 payment: payment,
-                package: package,
+                package: nil,
                 wrapper: self.storeKit1Wrapper
             ) { transaction, customerInfo, error, userCancelled in
                 continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
@@ -175,10 +276,27 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         expect(self.backend.invokedPostReceiptDataCount) == 0
     }
 
+    func testPurchaseSK1ReturnsMissingReceiptErrorIfSendReceiptFailed() async throws {
+        let product = try await self.fetchSk1Product()
+        let payment = self.storeKit1Wrapper.payment(with: product)
+        self.receiptFetcher.shouldReturnReceipt = false
+
+        let (_, _, error, _) = try await withCheckedThrowingContinuation { continuation in
+            self.orchestrator.purchase(
+                sk1Product: product,
+                payment: payment,
+                package: nil,
+                wrapper: self.storeKit1Wrapper
+            ) { transaction, customerInfo, error, userCancelled in
+                continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
+            }
+        }
+        expect(error).to(matchError(ErrorCode.missingReceiptFileError))
+    }
+
     // MARK: - Paywalls
 
-    func testPurchaseSK1PackageWithPresentedPaywall() async throws {
-        self.customerInfoManager.stubbedCachedCustomerInfoResult = self.mockCustomerInfo
+    func testPurchaseWithPresentedPaywall() async throws {
         self.backend.stubbedPostReceiptResult = .success(self.mockCustomerInfo)
 
         let product = try await self.fetchSk1Product()
@@ -205,7 +323,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         ) == Self.paywallEvent
     }
 
-    func testFailedSK1PurchaseRemembersPresentedPaywall() async throws {
+    func testPurchaseFailureRemembersPresentedPaywall() async throws {
         func purchase() async throws {
             let product = try await self.fetchSk1Product()
             let payment = self.storeKit1Wrapper.payment(with: product)
@@ -223,7 +341,6 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         }
 
         self.orchestrator.track(paywallEvent: .impression(Self.paywallEventCreationData, Self.paywallEvent))
-        self.customerInfoManager.stubbedCachedCustomerInfoResult = self.mockCustomerInfo
 
         self.backend.stubbedPostReceiptResult = .failure(.unexpectedBackendResponse(.customerInfoNil))
         try await purchase()
@@ -241,27 +358,17 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
 
     // MARK: - AdServices and Attributes
 
-    func testPurchaseSK1PackageDoesNotPostAdServicesTokenIfNotEnabled() async throws {
-        self.customerInfoManager.stubbedCachedCustomerInfoResult = self.mockCustomerInfo
+    func testPurchaseDoesNotPostAdServicesTokenIfNotEnabled() async throws {
         self.backend.stubbedPostReceiptResult = .success(self.mockCustomerInfo)
-
         self.attributionFetcher.adServicesTokenToReturn = "token"
-
         let product = try await self.fetchSk1Product()
-        let storeProduct = StoreProduct(sk1Product: product)
-
-        let package = Package(identifier: "package",
-                              packageType: .monthly,
-                              storeProduct: storeProduct,
-                              offeringIdentifier: "offering")
-
         let payment = self.storeKit1Wrapper.payment(with: product)
 
         _ = await withCheckedContinuation { continuation in
             self.orchestrator.purchase(
                 sk1Product: product,
                 payment: payment,
-                package: package,
+                package: nil,
                 wrapper: self.storeKit1Wrapper
             ) { transaction, customerInfo, error, userCancelled in
                 continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
@@ -274,13 +381,17 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
     @available(iOS 14.3, macOS 11.1, macCatalyst 14.3, *)
     @available(tvOS, unavailable)
     @available(watchOS, unavailable)
-    func testPurchaseSK1PackageWithSubscriberAttributesAndAdServicesToken() async throws {
+    func testPurchasePostsAdServicesTokenAndSubscriberAttributes() async throws {
         try AvailabilityChecks.skipIfTVOrWatchOSOrMacOS()
         try AvailabilityChecks.iOS14_3APIAvailableOrSkipTest()
 
         // Test for custom entitlement computation mode.
         // Without that mode, the token is posted upon calling `enableAdServicesAttributionTokenCollection`
-        self.systemInfo = .init(finishTransactions: true, customEntitlementsComputation: true)
+        self.systemInfo = .init(
+            finishTransactions: true,
+            customEntitlementsComputation: true,
+            storeKitVersion: .storeKit1
+        )
         self.setUpAttribution()
         self.setUpOrchestrator()
 
@@ -290,20 +401,11 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
             "attribute_2": .init(attribute: .email, value: "email")
         ]
 
-        self.customerInfoManager.stubbedCachedCustomerInfoResult = self.mockCustomerInfo
         self.backend.stubbedPostReceiptResult = .success(self.mockCustomerInfo)
-
         self.attributionFetcher.adServicesTokenToReturn = "token"
         self.subscriberAttributesManager.stubbedUnsyncedAttributesByKeyResult = attributes
         self.attribution.enableAdServicesAttributionTokenCollection()
-
         let product = try await self.fetchSk1Product()
-        let storeProduct = StoreProduct(sk1Product: product)
-
-        let package = Package(identifier: "package",
-                              packageType: .monthly,
-                              storeProduct: storeProduct,
-                              offeringIdentifier: "offering")
 
         let payment = self.storeKit1Wrapper.payment(with: product)
 
@@ -311,7 +413,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
             self.orchestrator.purchase(
                 sk1Product: product,
                 payment: payment,
-                package: package,
+                package: nil,
                 wrapper: self.storeKit1Wrapper
             ) { transaction, customerInfo, error, userCancelled in
                 continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
@@ -326,13 +428,11 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
 
     // MARK: - Promotional Offers
 
-    func testGetSK1PromotionalOffer() async throws {
-        customerInfoManager.stubbedCachedCustomerInfoResult = mockCustomerInfo
+    func testGetPromotionalOfferWorksIfThereIsATransaction() async throws {
         offerings.stubbedPostOfferCompletionResult = .success(("signature", "identifier", UUID(), 12345))
         self.receiptParser.stubbedReceiptHasTransactionsResult = true
 
         let product = try await fetchSk1Product()
-
         let storeProductDiscount = MockStoreProductDiscount(offerIdentifier: "offerid1",
                                                             currencyCode: product.priceLocale.currencyCode,
                                                             price: 11.1,
@@ -356,7 +456,8 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
             self.receiptFetcher.mockReceiptData.asFetchToken
     }
 
-    func testGetSK1PromotionalOfferFailsWithIneligibleIfNoReceiptIsFound() async throws {
+    func testGetPromotionalOfferSK1FailsWithIneligibleIfNoReceiptIsFound() async throws {
+        self.receiptParser.stubbedReceiptHasTransactionsResult = true
         self.receiptFetcher.shouldReturnReceipt = false
 
         let product = try await self.fetchSk1Product()
@@ -375,6 +476,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
                                                    product: StoreProduct(sk1Product: product),
                                                    completion: completion)
             }
+            fail("Expected error")
         } catch {
             expect(error).to(matchError(ErrorCode.ineligibleError))
         }
@@ -382,7 +484,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         expect(self.offerings.invokedPostOffer) == false
     }
 
-    func testGetSK1PromotionalOfferFailsWithIneligibleIfReceiptHasNoTransactions() async throws {
+    func testGetPromotionalOfferFailsWithIneligibleIfNoTransactionIsFound() async throws {
         self.receiptParser.stubbedReceiptHasTransactionsResult = false
 
         let product = try await self.fetchSk1Product()
@@ -401,6 +503,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
                                                    product: StoreProduct(sk1Product: product),
                                                    completion: completion)
             }
+            fail("Expected error")
         } catch {
             expect(error).to(matchError(ErrorCode.ineligibleError))
         }
@@ -408,35 +511,7 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         expect(self.offerings.invokedPostOffer) == false
     }
 
-    func testGetSK1PromotionalOfferWorksWhenReceiptHasTransactions() async throws {
-        customerInfoManager.stubbedCachedCustomerInfoResult = mockCustomerInfo
-        offerings.stubbedPostOfferCompletionResult = .success(("signature", "identifier", UUID(), 12345))
-        self.receiptParser.stubbedReceiptHasTransactionsResult = true
-
-        let product = try await self.fetchSk1Product()
-        let storeProductDiscount = MockStoreProductDiscount(offerIdentifier: "offerid1",
-                                                            currencyCode: product.priceLocale.currencyCode,
-                                                            price: 11.1,
-                                                            localizedPriceString: "$11.10",
-                                                            paymentMode: .payAsYouGo,
-                                                            subscriptionPeriod: .init(value: 1, unit: .month),
-                                                            numberOfPeriods: 2,
-                                                            type: .promotional)
-
-        let result = try await Async.call { completion in
-            orchestrator.promotionalOffer(forProductDiscount: storeProductDiscount,
-                                          product: StoreProduct(sk1Product: product),
-                                          completion: completion)
-        }
-
-        expect(result.signedData.identifier) == storeProductDiscount.offerIdentifier
-
-        expect(self.offerings.invokedPostOfferCount) == 1
-        expect(self.offerings.invokedPostOfferParameters?.offerIdentifier) == storeProductDiscount.offerIdentifier
-    }
-
-    func testGetSK1PromotionalOfferFailsWithIneligibleDiscount() async throws {
-        self.customerInfoManager.stubbedCachedCustomerInfoResult = mockCustomerInfo
+    func testGetPromotionalOfferFailsWithIneligibleIfBackendReturnsIneligible() async throws {
         self.offerings.stubbedPostOfferCompletionResult = .failure(
             .networkError(
                 .errorResponse(
@@ -464,7 +539,6 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
                                                    product: StoreProduct(sk1Product: product),
                                                    completion: completion)
             }
-
             fail("Expected error")
         } catch let purchasesError as PurchasesError {
             expect(purchasesError.error).to(matchError(ErrorCode.ineligibleError))
@@ -473,36 +547,92 @@ class PurchasesOrchestratorSK1Tests: PurchasesOrchestratorTestCase {
         }
     }
 
-    func testPurchaseSK1PackageWithDiscountSendsReceiptToBackendIfSuccessful() async throws {
-        customerInfoManager.stubbedCachedCustomerInfoResult = mockCustomerInfo
-        offerings.stubbedPostOfferCompletionResult = .success(("signature", "identifier", UUID(), 12345))
-        backend.stubbedPostReceiptResult = .success(mockCustomerInfo)
+    // MARK: - TransactionListenerDelegate
 
-        let product = try await fetchSk1Product()
-        let storeProduct = StoreProduct(sk1Product: product)
-        let package = Package(identifier: "package",
-                              packageType: .monthly,
-                              storeProduct: storeProduct,
-                              offeringIdentifier: "offering")
+    @available(iOS 15.0, tvOS 15.0, watchOS 8.0, macOS 12.0, *)
+    func testDoesNotListenForSK2TransactionsWithSK2Disabled() throws {
+        try AvailabilityChecks.iOS15APIAvailableOrSkipTest()
 
-        let offer = PromotionalOffer.SignedData(identifier: "",
-                                                keyIdentifier: "",
-                                                nonce: UUID(),
-                                                signature: "",
-                                                timestamp: 0)
+        let transactionListener = MockStoreKit2TransactionListener()
 
-        _ = await withCheckedContinuation { continuation in
-            orchestrator.purchase(sk1Product: product,
-                                  promotionalOffer: offer,
-                                  package: package,
-                                  wrapper: self.storeKit1Wrapper) { transaction, customerInfo, error, userCancelled in
-                continuation.resume(returning: (transaction, customerInfo, error, userCancelled))
-            }
+        self.setUpOrchestrator(storeKit2TransactionListener: transactionListener,
+                               storeKit2StorefrontListener: StoreKit2StorefrontListener(delegate: nil))
+
+        expect(transactionListener.invokedDelegateSetter).toEventually(beTrue())
+        expect(transactionListener.invokedListenForTransactions) == false
+    }
+
+    // MARK: - Sync Purchases
+
+    func testSyncPurchasesPostsReceipt() async throws {
+        self.backend.stubbedPostReceiptResult = .success(mockCustomerInfo)
+
+        let customerInfo = try await self.orchestrator.syncPurchases(receiptRefreshPolicy: .always,
+                                                                     isRestore: false,
+                                                                     initiationSource: .purchase)
+
+        expect(self.backend.invokedPostReceiptData).to(beTrue())
+        expect(self.backend.invokedPostReceiptDataParameters?.data) == .receipt(self.receiptFetcher.mockReceiptData)
+        expect(customerInfo) == mockCustomerInfo
+    }
+
+    func testSyncPurchasesDoesntPostReceiptAndReturnsCustomerInfoIfNoTransaction() async throws {
+        self.customerInfoManager.stubbedCachedCustomerInfoResult = mockCustomerInfo
+        self.receiptParser.stubbedReceiptHasTransactionsResult = false
+
+        let customerInfo = try await self.orchestrator.syncPurchases(receiptRefreshPolicy: .always,
+                                                                     isRestore: false,
+                                                                     initiationSource: .purchase)
+        expect(self.backend.invokedPostReceiptData).to(beFalse())
+        expect(self.customerInfoManager.invokedCustomerInfo).to(beFalse())
+        expect(customerInfo) == mockCustomerInfo
+    }
+
+    func testSyncPurchasesPostsReceiptIfNoTransactionsAndEmptyOriginalPurchaseDate() async throws {
+        self.customerInfoManager.stubbedCachedCustomerInfoResult = try .init(data: [
+            "request_date": "2019-08-16T10:30:42Z",
+            "subscriber": [
+                "first_seen": "2019-07-17T00:05:54Z",
+                "original_app_user_id": "app_user_id",
+                "subscriptions": [:] as [String: Any],
+                "other_purchases": [:] as [String: Any],
+                "original_application_version": "1.0"
+            ] as [String: Any]
+        ])
+        self.backend.stubbedPostReceiptResult = .success(self.mockCustomerInfo)
+        self.receiptParser.stubbedReceiptHasTransactionsResult = false
+
+        let customerInfo = try await self.orchestrator.syncPurchases(receiptRefreshPolicy: .always,
+                                                                     isRestore: false,
+                                                                     initiationSource: .purchase)
+        expect(self.backend.invokedPostReceiptData).to(beTrue())
+        expect(self.customerInfoManager.invokedCustomerInfo).to(beFalse())
+        expect(customerInfo) == mockCustomerInfo
+    }
+
+    func testSyncPurchasesCallsSuccessDelegateMethod() async throws {
+        self.backend.stubbedPostReceiptResult = .success(mockCustomerInfo)
+
+        let receivedCustomerInfo = try await self.orchestrator.syncPurchases(receiptRefreshPolicy: .always,
+                                                                             isRestore: false,
+                                                                             initiationSource: .purchase)
+
+        expect(receivedCustomerInfo) === mockCustomerInfo
+    }
+
+    func testSyncPurchasesPassesErrorOnFailure() async throws {
+        let expectedError: BackendError = .missingAppUserID()
+
+        self.backend.stubbedPostReceiptResult = .failure(expectedError)
+
+        do {
+            _ = try await self.orchestrator.syncPurchases(receiptRefreshPolicy: .always,
+                                                          isRestore: false,
+                                                          initiationSource: .purchase)
+            fail("Expected error")
+        } catch {
+            expect(error).to(matchError(expectedError.asPurchasesError))
         }
-
-        expect(self.backend.invokedPostReceiptDataCount) == 1
-        expect(self.backend.invokedPostReceiptDataParameters?.productData).toNot(beNil())
-        expect(self.backend.invokedPostReceiptDataParameters?.transactionData.presentedOfferingID) == "offering"
     }
 
 }
